refactor: User2 모델, 맵퍼, 수정
