+++ lib/mapper/user2_mapper.dart
@@ -1,11 +1,14 @@
 import 'package:modu_3_dart_study/dto/user2/user2_dto.dart';
-import 'package:modu_3_dart_study/model/user/user2.dart';
+import 'package:modu_3_dart_study/model/user2/user2.dart';
 
 extension User2DtoToModel on User2Dto {
   User2 toModel() {
     return User2(
-      id: id ?? 0,
+      id: id ?? User2.invalidId,
       name: name ?? '',
+      age: age ?? 0,
+      address: address ?? '',
+      phoneNumber: phoneNumber ?? '',
     );
   }
 }
@@ -15,6 +18,9 @@ extension User2ModelToDto on User2 {
     return User2Dto(
       id: id,
       name: name,
+      age: age,
+      address: address,
+      phoneNumber: phoneNumber,
     );
   }
 }
