+++ lib/repository/user2/user2_repository.dart
@@ -1,6 +1,6 @@
 import 'package:modu_3_dart_study/core/network_error.dart';
 import 'package:modu_3_dart_study/core/result.dart';
-import 'package:modu_3_dart_study/model/user/user2.dart';
+import 'package:modu_3_dart_study/model/user2/user2.dart';
 
 abstract interface class User2Repository {
   Future<Result<User2, NetworkError>> getUser(int id);
